+++ src/module/sheets/official/CharacterSheet.ts
@@ -376,7 +376,7 @@ export default class CharacterSheet extends ActorSheet {
       ...data.itemsByType['armor'],
       ...data.itemsByType['shield'],
     ]);
-    data.maxCarryCapacity = this._calcMaxCarryCapacity();
+    data.maxCarryCapacity = this.actor.calcMaxCarryCapacity();
 
     //Checks if the Actor has an Arcane Background
     data.hasArcaneBackground = this.actor.hasArcaneBackground;
@@ -474,20 +474,6 @@ export default class CharacterSheet extends ActorSheet {
       retVal += i.data.weight * i.data.quantity;
     });
     return retVal;
-  }
-
-  private _calcMaxCarryCapacity(): number {
-    const strengthDie = getProperty(
-      this.actor.data,
-      'data.attributes.strength.die',
-    );
-    let capacity = 20 + 10 * (strengthDie.sides - 4);
-
-    if (strengthDie.modifier > 0) {
-      capacity = capacity + 20 * strengthDie.modifier;
-    }
-
-    return capacity;
   }
 
   private _toggleEquipped(id: string, item: any): any {